chore(extension/podman): rename file having hyperv to hyper-v
