+++ extensions/podman/packages/extension/src/platforms/win-platform.spec.ts
@@ -20,8 +20,8 @@ import type { CheckResult, ExtensionContext, TelemetryLogger } from '@podman-des
 import * as extensionApi from '@podman-desktop/api';
 import { beforeEach, expect, test, vi } from 'vitest';
 
-import type { HyperVCheck } from '/@/checks/windows/hyperv-check';
-import type { HyperVPodmanVersionCheck } from '/@/checks/windows/hyperv-podman-version-check';
+import type { HyperVCheck } from '/@/checks/windows/hyper-v-check';
+import type { HyperVPodmanVersionCheck } from '/@/checks/windows/hyper-v-podman-version-check';
 import type { VirtualMachinePlatformCheck } from '/@/checks/windows/virtual-machine-platform-check';
 import type { WinBitCheck } from '/@/checks/windows/win-bit-check';
 import type { WinMemoryCheck } from '/@/checks/windows/win-memory-check';
